Streaming api moved to a new home under staging/cri-api/pkg
Refer to https://github.com/kubernetes/kubernetes/issues/75828#issue-426475237

Moving streaming under staging/cri-api/pkg as this library is indenpdent
to cri-api and is more of a helper library that can be reused for a
different use case as well.
