+++ staging/src/k8s.io/cri-api/BUILD
@@ -10,6 +10,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//staging/src/k8s.io/cri-api/pkg/apis:all-srcs",
+        "//staging/src/k8s.io/cri-api/pkg/streaming:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
