+++ test/test_owners.csv
@@ -664,7 +664,7 @@ k8s.io/kubernetes/pkg/kubelet/secret,kevin-wangzefeng,1,
 k8s.io/kubernetes/pkg/kubelet/server,tallclair,0,
 k8s.io/kubernetes/pkg/kubelet/server/portforward,rkouj,0,
 k8s.io/kubernetes/pkg/kubelet/server/stats,tallclair,0,
-k8s.io/kubernetes/pkg/kubelet/server/streaming,caesarxuchao,1,
+k8s.io/cri-api/pkg/streaming,caesarxuchao,1,
 k8s.io/kubernetes/pkg/kubelet/status,mwielgus,1,
 k8s.io/kubernetes/pkg/kubelet/sysctl,piosz,1,
 k8s.io/kubernetes/pkg/kubelet/types,jlowdermilk,1,
